requirements.txt
openai-whisper
gradio
scipy
numpy